feature(pauseScreen): moving structs to individual files
